fix: display api keys data
